yuna_Modify: User와 Alarm JPA 연관관계 적용에 따른 코드 수정

+++ backend/travel-on/src/main/java/com/travel/travel_on/controller/AlarmController.java
@@ -32,6 +32,8 @@ public class AlarmController {
     public ResponseEntity<?> selectAlarm(@PathVariable String id) {
         try {
             User user = usvc.select(id);
+            if(user==null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
             List<Alarm> list = asvc.selectAll(user);
             List<AlarmDto> result = list.stream()
                     .map(r -> new AlarmDto(r))
@@ -48,6 +50,8 @@ public class AlarmController {
     public ResponseEntity<?> deleteAlarm(@PathVariable String id) {
         try {
             User user = usvc.select(id);
+            if(user==null) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
             int result = asvc.deleteAll(user);
             return new ResponseEntity<Integer>(result, HttpStatus.OK);
         } catch (Exception e) {
